(refactor) split WebSocketManager god class: extract candle and Hummingbot stream families into condor/web/streams
