+++ condor/web/ws_manager.py
@@ -1,18 +1,35 @@
+"""WebSocket connection registry and channel broadcaster for the dashboard.
+
+The upstream stream implementations live in ``condor.web.streams`` (candles in
+``streams.candles``, the Hummingbot-bridged protocols in
+``streams.hummingbot_ws``) as mixins over the small host surface this module
+implements — see ``condor.web.streams.StreamHost``. ``WebSocketManager`` keeps
+only connection/subscription bookkeeping, the SDS bridge, broadcasting and the
+stream lifecycle registry.
+"""
+
 from __future__ import annotations
 
 import asyncio
 import json
 import logging
-import math
 import time
-from typing import TYPE_CHECKING, Any, Awaitable, Callable, Optional
+from typing import TYPE_CHECKING, Any, Optional
 
 from fastapi import WebSocket
 
-from condor import dex_candles
+# Re-exported for back-compat: tests and callers import the candle helpers
+# (and patch ``ws_manager.dex_candles``) through this module.
+from condor import dex_candles  # noqa: F401
 from condor.asyncutil import TaskSet
-from condor.fetchers.market_data import fetch_historical_candles, normalize_candle
 from condor.web.auth import decode_jwt
+from condor.web.streams.candles import (  # noqa: F401
+    CandleStreamsMixin,
+    _CandleBuffer,
+    _coerce_duration,
+    parse_candle_channel,
+)
+from condor.web.streams.hummingbot_ws import HummingbotStreamsMixin
 
 if TYPE_CHECKING:
     from condor.server_data_service import CacheKey
@@ -73,135 +90,6 @@ def channel_for_key(key: CacheKey) -> str | None:
     return ":".join([prefix, key.server, *segments])
 
 
-# Interval string -> seconds for buffer sizing
-_INTERVAL_SECONDS: dict[str, int] = {
-    "1s": 1,
-    "5s": 5,
-    "15s": 15,
-    "30s": 30,
-    "1m": 60,
-    "3m": 180,
-    "5m": 300,
-    "15m": 900,
-    "30m": 1800,
-    "1h": 3600,
-    "2h": 7200,
-    "4h": 14400,
-    "1d": 86400,
-    "1w": 604800,
-}
-
-# Auto-cleanup candle buffers unused for this long
-_CANDLE_BUFFER_IDLE_TTL = 600  # 10 minutes
-
-
-def _coerce_duration(value: object) -> int | None:
-    """Coerce a client-supplied candle duration to a positive int (seconds).
-
-    WS messages are untrusted input: a non-numeric value must not raise, or the
-    ValueError would escape handle_message and tear down the client's entire
-    multiplexed connection. Returns None for missing/invalid/non-positive values.
-    """
-    if value is None or isinstance(value, bool):
-        return None
-    try:
-        duration = int(float(value))
-    except (TypeError, ValueError):
-        logger.warning("Ignoring invalid candle duration from client: %r", value)
-        return None
-    return duration if duration > 0 else None
-
-
-def parse_candle_channel(channel: str) -> tuple[str, str, str, str, str | None] | None:
-    """Split a candle channel into its parts, or ``None`` if it is malformed.
-
-    The channel is ``candles:{server}:{connector}:{pair}:{interval}[:{pool_address}]``.
-    The pool address is an **optional sixth segment**, set by the DEX page so a
-    chart is pinned to the exact pool the user opened instead of whichever pool
-    ``dex_candles._resolve_pool`` judges deepest for the pair. A five-part channel
-    (every CLOB chart) parses to ``pool_address=None`` and is byte-identical to
-    what it was before the segment existed.
-
-    Every consumer must go through here: unpacking five names off ``split(":")``
-    raises ``ValueError`` on a six-part channel, and each of the three call sites
-    fails silently in a different way (a stream that never backfills, a backfill
-    that never ticks).
-    """
-    parts = channel.split(":")
-    if len(parts) < 5:
-        return None
-    server_name, connector, pair, interval = parts[1:5]
-    pool_address = parts[5] if len(parts) > 5 else None
-    return server_name, connector, pair, interval, pool_address or None
-
-
-class _CandleBuffer:
-    """Per-channel candle buffer with dynamic sizing based on interval + duration."""
-
-    __slots__ = ("interval", "_data", "_max_size", "last_accessed")
-
-    def __init__(self, interval: str, duration_seconds: int = 3600):
-        self.interval = interval
-        self._data: dict[float, dict] = {}
-        self._max_size: int = 200
-        self.last_accessed: float = time.monotonic()
-        self.set_duration(duration_seconds)
-
-    def set_duration(self, duration_seconds: int) -> int:
-        """Resize buffer for the given duration. Returns the new max size."""
-        interval_sec = _INTERVAL_SECONDS.get(self.interval, 60)
-        needed = math.ceil(duration_seconds / interval_sec)
-        new_max = max(needed, 200)  # minimum 200
-        old_max = self._max_size
-        self._max_size = new_max
-        self.last_accessed = time.monotonic()
-        self._evict()
-        if new_max != old_max:
-            logger.debug(
-                "Candle buffer resized %s: %d -> %d (duration=%ds)",
-                self.interval,
-                old_max,
-                new_max,
-                duration_seconds,
-            )
-        return new_max
-
-    def upsert(self, candle: dict) -> None:
-        self._data[candle["timestamp"]] = candle
-        self._evict()
-
-    def upsert_many(self, candles: list[dict]) -> None:
-        for c in candles:
-            self._data[c["timestamp"]] = c
-        self._evict()
-
-    def get_sorted(self) -> list[dict]:
-        self.last_accessed = time.monotonic()
-        return sorted(self._data.values(), key=lambda c: c["timestamp"])
-
-    def _evict(self) -> None:
-        excess = len(self._data) - self._max_size
-        if excess <= 0:
-            return
-        # Drop the `excess` oldest timestamps in a single O(n log n) pass
-        # instead of calling min() (O(n)) per evicted candle.
-        for ts in sorted(self._data)[:excess]:
-            del self._data[ts]
-
-    @property
-    def size(self) -> int:
-        return len(self._data)
-
-    @property
-    def max_size(self) -> int:
-        return self._max_size
-
-    @property
-    def needs_backfill(self) -> bool:
-        """True if buffer has room for significantly more candles."""
-        return len(self._data) < self._max_size * 0.5
-
-
 class _Connection:
     __slots__ = ("ws", "user_id", "channels")
 
@@ -211,15 +99,16 @@ class _Connection:
         self.channels: set[str] = set()
 
 
-class WebSocketManager:
+class WebSocketManager(CandleStreamsMixin, HummingbotStreamsMixin):
     """Manages WebSocket connections and channel-based data broadcasting.
 
     Subscribes to ServerDataService for data updates and broadcasts
-    to connected WebSocket clients. Candle streaming uses dedicated
-    WebSocket connections with dynamic per-channel buffering.
+    to connected WebSocket clients. The upstream stream protocols are
+    supplied by the mixins (``CandleStreamsMixin``,
+    ``HummingbotStreamsMixin``); this class owns their shared state and
+    the host surface they run on (``broadcast``, ``_send``,
+    ``_has_subscribers``, ``_ensure_stream``, ``_oneshot_tasks``).
     """
-
-    _CANDLE_KEEP_ALIVE = 300  # 5-minute grace period before tearing down candle streams
 
     def __init__(self):
         self._connections: list[_Connection] = []
@@ -263,11 +152,6 @@ class WebSocketManager:
         server-scoped; returns None when no server segment is present."""
         parts = channel.split(":")
         return parts[1] if len(parts) >= 2 and parts[1] else None
-
-    @staticmethod
-    def _normalize_candle(c: Any) -> dict | None:
-        """Normalize a candle from any format to a uniform dict with float values."""
-        return normalize_candle(c)
 
     # -- Lifecycle --
 
@@ -450,135 +334,7 @@ class WebSocketManager:
             if channel.startswith("candles:") and duration:
                 await self._handle_candle_duration_change(conn, channel, duration)
 
-    async def _handle_candle_subscribe(
-        self, conn: _Connection, channel: str, duration: object
-    ) -> None:
-        """Handle candle channel subscription with optional duration."""
-        parts = channel.split(":")
-        if len(parts) < 5:
-            return
-        interval = parts[4]
-        dur = _coerce_duration(duration) or 3 * 86400  # default 3 days
-
-        # Cancel any pending teardown — a subscriber just came back
-        timer = self._candle_teardown_timers.pop(channel, None)
-        if timer is not None:
-            timer.cancel()
-            logger.debug("Cancelled pending candle teardown for %s", channel)
-
-        buf = self._candle_buffers.get(channel)
-        if buf is None:
-            buf = _CandleBuffer(interval, dur)
-            self._candle_buffers[channel] = buf
-        else:
-            # Expand buffer if this client needs more
-            if dur > 0:
-                old_max = buf.max_size
-                buf.set_duration(dur)
-                if buf.max_size > old_max and buf.needs_backfill:
-                    self._oneshot_tasks.track(
-                        asyncio.create_task(
-                            self._backfill_candles(channel),
-                            name=f"backfill:{channel}",
-                        )
-                    )
-
-        # Send buffered candles as initial snapshot
-        sorted_candles = buf.get_sorted()
-        if sorted_candles:
-            await self._send(conn, channel, {"type": "candles", "data": sorted_candles})
-
-        # If the stream task is still running (kept alive during grace period), skip restart
-        self._ensure_stream("candles", channel)
-
-    async def _handle_candle_duration_change(
-        self, conn: _Connection, channel: str, duration: int
-    ) -> None:
-        """Handle duration change for an existing candle subscription.
-
-        Only grows the buffer — never shrinks. The frontend manages its own
-        display window; the backend just ensures enough history is buffered.
-        """
-        buf = self._candle_buffers.get(channel)
-        if buf is None:
-            return
-        old_max = buf.max_size
-        # Only expand — skip if requested duration would shrink the buffer
-        interval_sec = _INTERVAL_SECONDS.get(buf.interval, 60)
-        needed = max(math.ceil(duration / interval_sec), 200)
-        if needed <= old_max:
-            return
-        buf.set_duration(duration)
-        if buf.needs_backfill:
-            await self._backfill_candles(channel)
-        # Broadcast updated snapshot to ALL subscribers on this channel
-        sorted_candles = buf.get_sorted()
-        if sorted_candles:
-            await self.broadcast(channel, {"type": "candles", "data": sorted_candles})
-
-    async def _backfill_candles(self, channel: str) -> None:
-        """Fetch historical candles to fill the buffer gap."""
-        parsed = parse_candle_channel(channel)
-        if parsed is None:
-            return
-        server_name, connector, pair, interval, pool_address = parsed
-        buf = self._candle_buffers.get(channel)
-        if buf is None:
-            return
-
-        from config_manager import get_config_manager
-
-        cm = get_config_manager()
-        try:
-            interval_sec = _INTERVAL_SECONDS.get(interval, 60)
-            end_time = int(time.time())
-            start_time = end_time - (buf.max_size * interval_sec)
-
-            logger.info(
-                "Backfilling candles for %s: need %d, have %d, fetching %ds-%ds",
-                channel,
-                buf.max_size,
-                buf.size,
-                start_time,
-                end_time,
-            )
-
-            if dex_candles.uses_gecko_candles(connector):
-                candles = await dex_candles.fetch_dex_candles(
-                    connector, pool_address, pair, interval, start_time, end_time
-                )
-                if candles:
-                    buf.upsert_many(candles)
-                    logger.info(
-                        "Backfilled %d GeckoTerminal candles for %s (buffer: %d/%d)",
-                        len(candles),
-                        channel,
-                        buf.size,
-                        buf.max_size,
-                    )
-                return
-
-            client = await cm.get_client(server_name)
-            candles = await fetch_historical_candles(
-                client,
-                connector,
-                pair,
-                interval,
-                start_time=start_time,
-                end_time=end_time,
-                limit=min(buf.max_size, 5000),
-            )
-            if candles:
-                buf.upsert_many(candles)
-                logger.info(
-                    "Backfilled %d candles for %s (buffer: %d/%d)",
-                    len(candles),
-                    channel,
-                    buf.size,
-                    buf.max_size,
-                )
-        except Exception as e:
-            logger.warning("Candle backfill failed for %s: %s", channel, e)
+    # -- SDS bridge --
 
     async def _subscribe_sds(self, channel: str) -> None:
         """Subscribe to SDS for a channel and prime cache."""
@@ -698,36 +454,6 @@ class WebSocketManager:
             )
 
     # -- SDS listener --
-
-    @staticmethod
-    def _transform_executors(raw_data: Any) -> list[dict]:
-        """Transform raw executor data to ExecutorInfo-compatible dicts for WS broadcast."""
-        from condor.fetchers.executors import extract_executors_list
-        from condor.web.models import ExecutorInfo
-
-        executors_list = extract_executors_list(raw_data)
-        result = []
-        for ex in executors_list:
-            info = ExecutorInfo.from_raw(ex)
-            if info:
-                result.append(info.model_dump())
-        return result
-
-    @staticmethod
-    def _transform_bots(raw_data: Any) -> dict:
-        """Transform raw BOTS_STATUS data to BotsPageResponse-compatible dict for WS broadcast."""
-        from condor.fetchers.bots import build_bots_page
-
-        return build_bots_page(raw_data)
-
-    @staticmethod
-    def _overlay_stopping_state(server_name: str, data: dict) -> None:
-        """Apply transitional 'stopping' state to WS broadcast data."""
-        from condor.web.routes.bots import overlay_stopping_state
-
-        overlay_stopping_state(
-            server_name, data.get("controllers", []), data.get("bots", [])
-        )
 
     def _on_data_update(self, key: CacheKey, value: Any) -> None:
         """Called by SDS on every cache write. Maps the key to a WS channel and
@@ -890,913 +616,6 @@ class WebSocketManager:
             task.cancel()
             logger.info(spec["stop_log"], channel)
         self._last_data.pop(channel, None)
-
-    def _maybe_stop_candle_stream(self, channel: str) -> None:
-        # If subscribers still exist, cancel any pending teardown and return
-        if self._has_subscribers(channel):
-            timer = self._candle_teardown_timers.pop(channel, None)
-            if timer is not None:
-                timer.cancel()
-            return
-
-        # Already have a pending teardown scheduled — nothing to do
-        if channel in self._candle_teardown_timers:
-            return
-
-        # Schedule deferred teardown
-        loop = asyncio.get_event_loop()
-        handle = loop.call_later(
-            self._CANDLE_KEEP_ALIVE,
-            self._deferred_stop_candle_stream,
-            channel,
-        )
-        self._candle_teardown_timers[channel] = handle
-        logger.info(
-            "Scheduled candle stream teardown for %s in %ds",
-            channel,
-            self._CANDLE_KEEP_ALIVE,
-        )
-
-    def _deferred_stop_candle_stream(self, channel: str) -> None:
-        """Actually tear down a candle stream after the grace period."""
-        self._candle_teardown_timers.pop(channel, None)
-
-        # Re-check: subscribers may have appeared during the grace period
-        if self._has_subscribers(channel):
-            logger.debug(
-                "Candle teardown cancelled — subscribers returned for %s", channel
-            )
-            return
-
-        task = self._candle_tasks.pop(channel, None)
-        if task and not task.done():
-            task.cancel()
-            logger.info("Deferred candle stream teardown completed for %s", channel)
-        poll_task = self._candle_poll_tasks.pop(channel, None)
-        if poll_task and not poll_task.done():
-            poll_task.cancel()
-        self._last_candle_ws_update.pop(channel, None)
-        self._candle_first_msg_logged.discard(channel)
-        # Snapshot-on-resubscribe is served by _candle_buffers, not _last_data
-        self._last_data.pop(channel, None)
-        # NOTE: candle buffer is NOT deleted — the existing idle cleanup loop handles that
-
-    @staticmethod
-    def _is_permanent_ws_error(error_str: str) -> bool:
-        """True if a stream error will never succeed on retry.
-
-        Besides auth/not-found (401/403/404), an invalid trading pair (wrong
-        symbol for this connector) will never become valid, so retrying
-        forever just spams logs and hammers the exchange until Condor
-        restarts (issue #134).
-        """
-        lowered = error_str.lower()
-        return (
-            any(code in error_str for code in ("401", "403", "404"))
-            or "appears to be invalid" in lowered
-            or "invalid symbol" in lowered
-        )
-
-    async def _run_ws_stream(
-        self,
-        channel: str,
-        server_name: str,
-        *,
-        label: str,
-        open_ws: Callable[[Any], Any],
-        subscribe: Callable[[Any], Awaitable[None]],
-        on_message: Callable[[dict], Awaitable[None]],
-        backoff_on_empty_close: bool = False,
-    ) -> None:
-        """Shared connect/reconnect skeleton for all Hummingbot WS streams.
-
-        Owns the while-True loop, subscriber check, heartbeat/error message
-        handling, permanent-error detection (``_is_permanent_ws_error``) and
-        exponential backoff capped at 60s. Each stream supplies only:
-
-        - ``open_ws``: client -> the WS async context manager to enter
-          (e.g. ``lambda c: c.ws.market_data()``).
-        - ``subscribe``: performs the subscription on the open socket.
-        - ``on_message``: handles data messages (heartbeat/error are
-          handled here).
-
-        ``backoff_on_empty_close`` reproduces the executor stream's
-        behavior: don't reset backoff on subscribe; after a clean close,
-        reset it only if at least one message arrived, otherwise back off
-        (guards against a server that accepts then immediately drops).
-        """
-        from config_manager import get_config_manager
-
-        cm = get_config_manager()
-        backoff = 5
-        lower_label = label[0].lower() + label[1:]
-        subscribed_label = label if label.endswith("WS") else f"{label} WS"
-
-        while True:
-            try:
-                client = await cm.get_client(server_name)
-                async with open_ws(client) as ws:
-                    await subscribe(ws)
-                    logger.info("%s subscribed: %s", subscribed_label, channel)
-                    if not backoff_on_empty_close:
-                        backoff = 5
-                    got_message = False
-                    async for msg in ws:
-                        if not self._has_subscribers(channel):
-                            logger.info(
-                                "No subscribers for %s, closing %s stream",
-                                channel,
-                                lower_label,
-                            )
-                            return
-
-                        got_message = True
-                        msg_type = msg.get("type")
-                        if msg_type == "heartbeat":
-                            continue
-                        if msg_type == "error":
-                            logger.warning(
-                                "%s stream error for %s: %s",
-                                label,
-                                channel,
-                                msg.get("message", "unknown error"),
-                            )
-                            break
-                        await on_message(msg)
-
-                    if backoff_on_empty_close:
-                        # Connection closed cleanly — back off if short-lived
-                        if got_message:
-                            backoff = 5
-                        else:
-                            logger.warning(
-                                "%s stream closed immediately for %s, "
-                                "reconnecting in %ds...",
-                                label,
-                                channel,
-                                backoff,
-                            )
-                            await asyncio.sleep(backoff)
-                            backoff = min(backoff * 2, 60)
-
-            except asyncio.CancelledError:
-                return
-            except Exception as e:
-                if self._is_permanent_ws_error(str(e)):
-                    logger.warning(
-                        "%s stream permanent error for %s: %s — giving up",
-                        label,
-                        channel,
-                        e,
-                    )
-                    return
-
-                logger.warning(
-                    "%s stream error for %s: %s, reconnecting in %ds...",
-                    label,
-                    channel,
-                    e,
-                    backoff,
-                )
-                await asyncio.sleep(backoff)
-                backoff = min(backoff * 2, 60)
-
-    async def _candle_stream(self, channel: str) -> None:
-        parsed = parse_candle_channel(channel)
-        if parsed is None:
-            return
-        server_name, connector, pair, interval, pool_address = parsed
-
-        from config_manager import get_config_manager
-
-        cm = get_config_manager()
-        backoff = 5
-
-        # Start REST poll fallback alongside the WS stream
-        self._ensure_candle_poll_fallback(channel)
-
-        if dex_candles.uses_gecko_candles(connector):
-            # No CandlesFactory feed to subscribe to (xrpl, DEX networks): the
-            # poll loop above is the whole stream. Subscribing anyway would just
-            # reconnect-loop on an error the API can never stop returning.
-            logger.info(
-                "Candle stream for %s is GeckoTerminal-polled (no API feed)", channel
-            )
-            return
-
-        while True:
-            try:
-                client = await cm.get_client(server_name)
-                async with client.ws.market_data() as ws:
-                    await ws.subscribe_candles(
-                        connector,
-                        pair,
-                        interval=interval,
-                        max_records=100,
-                        update_interval=1.0,
-                    )
-                    logger.info("Candle WS subscribed: %s", channel)
-                    backoff = 5  # Reset on successful connection
-                    async for msg in ws:
-                        # Log first message per channel at INFO for diagnostics
-                        if channel not in self._candle_first_msg_logged:
-                            self._candle_first_msg_logged.add(channel)
-                            logger.info(
-                                "Candle WS first message for %s: %s",
-                                channel,
-                                (
-                                    json.dumps(msg)[:500]
-                                    if isinstance(msg, dict)
-                                    else type(msg).__name__
-                                ),
-                            )
-                        else:
-                            logger.debug(
-                                "Candle WS raw msg keys for %s: %s",
-                                channel,
-                                (
-                                    list(msg.keys())
-                                    if isinstance(msg, dict)
-                                    else type(msg).__name__
-                                ),
-                            )
-
-                        msg_type = msg.get("type")
-                        if msg_type == "candle_update":
-                            raw = msg.get("data")
-                            candle = self._normalize_candle(raw) if raw else None
-                            if candle:
-                                self._last_candle_ws_update[channel] = time.monotonic()
-                                self._upsert_candle_buffer(channel, candle)
-                                await self.broadcast(
-                                    channel,
-                                    {"type": "candle_update", "candle": candle},
-                                )
-                        elif msg_type == "candles":
-                            raw_list = msg.get("data") or []
-                            candles = [
-                                c
-                                for r in raw_list
-                                if (c := self._normalize_candle(r)) is not None
-                            ]
-                            if candles:
-                                self._last_candle_ws_update[channel] = time.monotonic()
-                                self._upsert_candle_buffer_many(channel, candles)
-                                await self.broadcast(
-                                    channel,
-                                    {"type": "candles", "data": candles},
-                                )
-                        elif msg_type in ("heartbeat", "subscribed"):
-                            continue
-                        elif msg_type == "error":
-                            error_msg = msg.get("message", "unknown error")
-                            logger.warning(
-                                "Candle stream error for %s: %s — continuing",
-                                channel,
-                                error_msg,
-                            )
-                            await self.broadcast(
-                                channel,
-                                {
-                                    "type": "error",
-                                    "message": f"Stream error: {error_msg}",
-                                },
-                            )
-                            # Don't break — the WS may still be alive.
-                            # If truly dead, next recv raises and we reconnect.
-                            continue
-                        else:
-                            logger.info(
-                                "Candle stream unrecognized msg type for %s: type=%s keys=%s",
-                                channel,
-                                msg_type,
-                                (
-                                    list(msg.keys())
-                                    if isinstance(msg, dict)
-                                    else type(msg).__name__
-                                ),
-                            )
-
-            except asyncio.CancelledError:
-                return
-            except Exception as e:
-                error_str = str(e)
-                # Detect permanent failures — don't retry (issue #134, see
-                # _is_permanent_ws_error).
-                if self._is_permanent_ws_error(error_str):
-                    logger.warning(
-                        "Candle stream permanent error for %s: %s — giving up",
-                        channel,
-                        e,
-                    )
-                    await self.broadcast(
-                        channel,
-                        {"type": "error", "message": f"Stream failed: {error_str}"},
-                    )
-                    # Stop the paired REST poll fallback — nothing valid to poll.
-                    poll_task = self._candle_poll_tasks.pop(channel, None)
-                    if poll_task and not poll_task.done():
-                        poll_task.cancel()
-                    return
-
-                logger.warning(
-                    "Candle stream error for %s: %s, reconnecting in %ds...",
-                    channel,
-                    e,
-                    backoff,
-                )
-                await self.broadcast(
-                    channel,
-                    {
-                        "type": "error",
-                        "message": f"Connection lost, retrying in {backoff}s",
-                    },
-                )
-                await asyncio.sleep(backoff)
-                backoff = min(backoff * 2, 60)
-
-    # -- Candle REST poll fallback --
-
-    def _ensure_candle_poll_fallback(self, channel: str) -> None:
-        """Start a REST poll fallback task for a candle channel if not already running."""
-        if (
-            channel in self._candle_poll_tasks
-            and not self._candle_poll_tasks[channel].done()
-        ):
-            return
-        self._candle_poll_tasks[channel] = asyncio.create_task(
-            self._candle_poll_fallback(channel)
-        )
-
-    async def _candle_poll_fallback(self, channel: str) -> None:
-        """Periodically poll REST for candles when WS stream goes silent.
-
-        For a GeckoTerminal-backed connector there is no WS stream to fall back
-        from — this loop *is* the feed, so it polls from the first tick and never
-        waits for staleness.
-        """
-        parsed = parse_candle_channel(channel)
-        if parsed is None:
-            return
-        server_name, connector, pair, interval, pool_address = parsed
-        interval_sec = _INTERVAL_SECONDS.get(interval, 60)
-        gecko = dex_candles.uses_gecko_candles(connector)
-        # How long without a WS update before we consider it stale
-        stale_threshold = max(interval_sec, 15)
-        # How often to poll REST once stale (keep the last candle fresh). Gecko
-        # polls stay well clear of its per-minute rate limit — every chart on the
-        # dashboard shares one budget — while still refreshing a forming candle.
-        poll_interval = (
-            max(min(interval_sec // 2, 60), 30) if gecko else min(interval_sec, 10)
-        )
-        was_stale = gecko
-
-        from config_manager import get_config_manager
-
-        try:
-            while True:
-                await asyncio.sleep(poll_interval if was_stale else stale_threshold)
-
-                last_update = self._last_candle_ws_update.get(channel)
-                if (
-                    last_update is not None
-                    and (time.monotonic() - last_update) <= stale_threshold
-                ):
-                    if was_stale:
-                        logger.info(
-                            "Candle WS stream resumed for %s, stopping REST fallback polling",
-                            channel,
-                        )
-                        was_stale = False
-                    continue
-
-                # Stream is stale — poll REST
-                if not was_stale:
-                    logger.warning(
-                        "Candle WS stream appears stale for %s (threshold=%ds), polling REST fallback",
-                        channel,
-                        stale_threshold,
-                    )
-                    was_stale = True
-
-                try:
-                    now = int(time.time())
-                    if gecko:
-                        candles = await dex_candles.fetch_dex_candles(
-                            connector,
-                            pool_address,
-                            pair,
-                            interval,
-                            now - interval_sec * 5,
-                            now,
-                            # The shared OHLCV cache holds a pool for minutes,
-                            # long enough to freeze a live chart.
-                            use_cache=False,
-                        )
-                    else:
-                        cm = get_config_manager()
-                        client = await cm.get_client(server_name)
-                        # No `limit`: this poll only wants the fresh tail of the
-                        # live range, so an empty historical answer stays empty
-                        # rather than pulling a full unrelated window.
-                        candles = await fetch_historical_candles(
-                            client,
-                            connector,
-                            pair,
-                            interval,
-                            start_time=now - interval_sec * 5,
-                            end_time=now,
-                        )
-                    if candles:
-                        buf = self._candle_buffers.get(channel)
-                        # Broadcast if we have newer candles OR if the latest
-                        # candle's OHLCV changed (same timestamp, updated values)
-                        newest_buf_ts = (
-                            max(buf._data.keys()) if buf and buf._data else 0
-                        )
-                        newest_poll_ts = max(c["timestamp"] for c in candles)
-                        changed = newest_poll_ts > newest_buf_ts
-                        if not changed and buf and newest_poll_ts in buf._data:
-                            # Same timestamp — check if OHLCV actually changed
-                            old = buf._data[newest_poll_ts]
-                            new = next(
-                                c for c in candles if c["timestamp"] == newest_poll_ts
-                            )
-                            changed = any(
-                                old.get(k) != new.get(k)
-                                for k in ("open", "high", "low", "close", "volume")
-                            )
-                        if changed:
-                            self._upsert_candle_buffer_many(channel, candles)
-                            await self.broadcast(
-                                channel,
-                                {"type": "candles", "data": candles},
-                            )
-                            logger.debug(
-                                "REST fallback delivered %d candles for %s",
-                                len(candles),
-                                channel,
-                            )
-                except Exception as e:
-                    logger.debug("REST candle poll failed for %s: %s", channel, e)
-
-        except asyncio.CancelledError:
-            return
-
-    # -- Trade streaming --
-
-    async def _trade_stream(self, channel: str) -> None:
-        parts = channel.split(":")
-        if len(parts) < 4:
-            return
-        _, server_name, connector, pair = parts[:4]
-
-        async def subscribe(ws: Any) -> None:
-            await ws.subscribe_trades(
-                connector,
-                pair,
-                update_interval=1.0,
-            )
-
-        async def on_message(msg: dict) -> None:
-            if msg.get("type") != "trades":
-                return
-            trade_data = msg.get("data", [])
-            trades = []
-            for t in trade_data:
-                if isinstance(t, dict):
-                    trades.append(
-                        {
-                            "price": float(t.get("price", 0)),
-                            "amount": float(t.get("amount", t.get("quantity", 0))),
-                            "side": t.get("side", t.get("trade_type", "buy")).lower(),
-                            "timestamp": float(t.get("timestamp", 0)),
-                        }
-                    )
-            if trades:
-                await self.broadcast(channel, {"type": "trades", "data": trades})
-
-        await self._run_ws_stream(
-            channel,
-            server_name,
-            label="Trade",
-            open_ws=lambda client: client.ws.market_data(),
-            subscribe=subscribe,
-            on_message=on_message,
-        )
-
-    # -- Candle buffer helpers --
-
-    def _get_or_create_candle_buffer(self, channel: str) -> _CandleBuffer:
-        """Get existing buffer or create with default duration."""
-        buf = self._candle_buffers.get(channel)
-        if buf is None:
-            parts = channel.split(":")
-            interval = parts[4] if len(parts) >= 5 else "1m"
-            buf = _CandleBuffer(interval, 3 * 86400)  # default 3 days
-            self._candle_buffers[channel] = buf
-        return buf
-
-    def _upsert_candle_buffer(self, channel: str, candle: dict) -> None:
-        """Upsert a single candle into the per-channel buffer."""
-        buf = self._get_or_create_candle_buffer(channel)
-        buf.upsert(candle)
-
-    def _upsert_candle_buffer_many(self, channel: str, candles: list[dict]) -> None:
-        """Upsert multiple candles into the per-channel buffer."""
-        buf = self._get_or_create_candle_buffer(channel)
-        buf.upsert_many(candles)
-
-    async def _candle_buffer_cleanup_loop(self) -> None:
-        """Periodically remove candle buffers that haven't been accessed."""
-        try:
-            while True:
-                await asyncio.sleep(60)  # check every minute
-                now = time.monotonic()
-                stale = [
-                    ch
-                    for ch, buf in self._candle_buffers.items()
-                    if (now - buf.last_accessed) > _CANDLE_BUFFER_IDLE_TTL
-                    and not any(ch in c.channels for c in self._connections)
-                ]
-                for ch in stale:
-                    buf = self._candle_buffers.pop(ch, None)
-                    if buf:
-                        logger.info(
-                            "Cleaned up idle candle buffer: %s (%d candles)",
-                            ch,
-                            buf.size,
-                        )
-        except asyncio.CancelledError:
-            return
-
-    # -- Order book streaming --
-
-    async def _order_book_stream(self, channel: str) -> None:
-        parts = channel.split(":")
-        if len(parts) < 4:
-            return
-        _, server_name, connector, pair = parts[:4]
-
-        async def subscribe(ws: Any) -> None:
-            await ws.subscribe_order_book(
-                connector,
-                pair,
-                depth=20,
-                update_interval=1.0,
-            )
-
-        async def on_message(msg: dict) -> None:
-            if msg.get("type") != "order_book":
-                return
-            raw_data = msg.get("data", {})
-            bids = []
-            asks = []
-            for b in raw_data.get("bids") or []:
-                if isinstance(b, dict):
-                    bids.append(
-                        {
-                            "price": float(b.get("price", 0)),
-                            "amount": float(b.get("amount", b.get("quantity", 0))),
-                        }
-                    )
-                elif isinstance(b, (list, tuple)) and len(b) >= 2:
-                    bids.append({"price": float(b[0]), "amount": float(b[1])})
-            for a in raw_data.get("asks") or []:
-                if isinstance(a, dict):
-                    asks.append(
-                        {
-                            "price": float(a.get("price", 0)),
-                            "amount": float(a.get("amount", a.get("quantity", 0))),
-                        }
-                    )
-                elif isinstance(a, (list, tuple)) and len(a) >= 2:
-                    asks.append({"price": float(a[0]), "amount": float(a[1])})
-            await self.broadcast(channel, {"bids": bids, "asks": asks})
-
-        await self._run_ws_stream(
-            channel,
-            server_name,
-            label="Order book",
-            open_ws=lambda client: client.ws.market_data(),
-            subscribe=subscribe,
-            on_message=on_message,
-        )
-
-    # -- Executor streaming (via Hummingbot WS) --
-
-    async def _executor_stream(self, channel: str) -> None:
-        parts = channel.split(":")
-        if len(parts) < 2:
-            return
-        server_name = parts[1]
-
-        from config_manager import get_config_manager
-
-        cm = get_config_manager()
-
-        # Try SDS cache first (pre-warmed by auto_subscribe_servers or REST prefetch)
-        from condor.server_data_service import ServerDataType, get_server_data_service
-
-        if channel not in self._last_data:
-            sds = get_server_data_service()
-            cached = sds.get(server_name, ServerDataType.EXECUTORS)
-            if cached is not None:
-                executors = self._transform_executors(cached)
-                if executors:
-                    await self.broadcast(channel, executors)
-                    logger.info(
-                        "Executor SDS cache hit: %d executors for %s",
-                        len(executors),
-                        channel,
-                    )
-
-        # Wait briefly for SDS to be populated by a concurrent REST request
-        # (usePrefetchData fires getExecutors which calls get_or_fetch on SDS)
-        if channel not in self._last_data:
-            sds = get_server_data_service()
-            for _ in range(6):  # up to 3 seconds
-                await asyncio.sleep(0.5)
-                cached = sds.get(server_name, ServerDataType.EXECUTORS)
-                if cached is not None:
-                    executors = self._transform_executors(cached)
-                    if executors:
-                        await self.broadcast(channel, executors)
-                        logger.info(
-                            "Executor SDS cache populated during wait: %d executors for %s",
-                            len(executors),
-                            channel,
-                        )
-                    break
-
-        # Progressive pre-fetch only if we still have no data
-        if channel not in self._last_data:
-            try:
-                from condor.fetchers._pagination import walk_pages
-                from condor.fetchers.executors import (
-                    extract_executors_list as _extract_executors_list,
-                )
-
-                sds = get_server_data_service()
-                client = await cm.get_client(server_name)
-                all_raw: list[dict] = []
-                # Rows already through ExecutorInfo, accumulated page by page.
-                # Transforming the *whole* accumulated list on every page made
-                # this quadratic: with FIRST_PAGE/NEXT_PAGE/MAX_PREFETCH below,
-                # a 5k history cost ~27k pydantic validations instead of 5k, all
-                # on the event loop the candle and bots broadcasts share.
-                # ``_transform_executors`` is a pure per-item map over a list
-                # (``extract_executors_list`` passes a list through untouched),
-                # so transforming each page once yields the identical snapshot.
-                transformed: list[dict] = []
-                page_num = 0
-                FIRST_PAGE = 50
-                NEXT_PAGE = 500
-                MAX_PREFETCH = 5000
-
-                # A small first page so the tab paints before the long pages land.
-                async for page in walk_pages(
-                    client.executors.search_executors,
-                    _extract_executors_list,
-                    page_size=NEXT_PAGE,
-                    first_page_size=FIRST_PAGE,
-                    max_items=MAX_PREFETCH,
-                ):
-                    all_raw.extend(page)
-                    transformed.extend(self._transform_executors(page))
-
-                    # Broadcast the accumulated snapshot after each page. It is
-                    # copied because ``broadcast`` retains what it is handed as
-                    # the channel's last-known payload, which the next page
-                    # would otherwise keep mutating underneath it.
-                    if transformed:
-                        await self.broadcast(channel, list(transformed))
-                        logger.info(
-                            "Executor pre-fetch page %d: %d executors (total %d) for %s",
-                            page_num,
-                            len(page),
-                            len(transformed),
-                            channel,
-                        )
-                    page_num += 1
-
-                # Cache in SDS so other consumers benefit
-                sds.put(server_name, ServerDataType.EXECUTORS, all_raw)
-            except asyncio.CancelledError:
-                return
-            except Exception as e:
-                logger.warning("Executor pre-fetch failed for %s: %s", channel, e)
-
-        async def subscribe(ws: Any) -> None:
-            await ws.subscribe_executors(update_interval=2.0)
-
-        async def on_message(msg: dict) -> None:
-            if msg.get("type") != "executors":
-                return
-            executors = self._transform_executors(msg.get("data", []))
-            await self._broadcast_update(channel, executors)
-
-        await self._run_ws_stream(
-            channel,
-            server_name,
-            label="Executor",
-            open_ws=lambda client: client.ws.executors(),
-            subscribe=subscribe,
-            on_message=on_message,
-            backoff_on_empty_close=True,
-        )
-
-    # -- Bots WS streaming (via Hummingbot /ws/executors all_bots_status) --
-
-    async def _bots_ws_stream(self, channel: str) -> None:
-        """Stream all_bots_status from Hummingbot /ws/executors and update SDS cache."""
-        parts = channel.split(":")
-        if len(parts) < 2:
-            return
-        server_name = parts[1]
-
-        # Send SDS-cached bots data as initial snapshot
-        if channel not in self._last_data:
-            from condor.server_data_service import (
-                ServerDataType,
-                get_server_data_service,
-            )
-
-            sds = get_server_data_service()
-            cached = sds.get(server_name, ServerDataType.BOTS_STATUS)
-            if cached is not None:
-                try:
-                    data = self._transform_bots(cached)
-                    await self.broadcast(channel, data)
-                except Exception as e:
-                    logger.debug("Failed to send initial bots snapshot: %s", e)
-
-        async def subscribe(ws: Any) -> None:
-            # all_bots_status is not in the client library, send raw
-            await ws._send(
-                {
-                    "action": "subscribe",
-                    "type": "all_bots_status",
-                    "update_interval": 5.0,
-                }
-            )
-            resp = await ws._receive()
-            if resp.get("type") == "error":
-                raise RuntimeError(f"Subscribe failed: {resp.get('message')}")
-
-        async def on_message(msg: dict) -> None:
-            if msg.get("type") != "all_bots_status":
-                return
-            raw_data = msg.get("data", {})
-            # Update SDS cache so REST and Telegram benefit
-            from condor.server_data_service import (
-                ServerDataType,
-                get_server_data_service,
-            )
-
-            get_server_data_service().put(
-                server_name, ServerDataType.BOTS_STATUS, raw_data
-            )
-            try:
-                data = self._transform_bots(raw_data)
-                self._overlay_stopping_state(server_name, data)
-                await self._broadcast_update(channel, data)
-            except Exception as e:
-                logger.debug("Failed to transform bots WS data: %s", e)
-
-        await self._run_ws_stream(
-            channel,
-            server_name,
-            label="Bots WS",
-            open_ws=lambda client: client.ws.executors(),
-            subscribe=subscribe,
-            on_message=on_message,
-        )
-
-    # -- Positions WS streaming (via Hummingbot /ws/executors positions) --
-
-    async def _positions_ws_stream(self, channel: str) -> None:
-        """Stream positions from Hummingbot /ws/executors and update SDS cache."""
-        parts = channel.split(":")
-        if len(parts) < 2:
-            return
-        server_name = parts[1]
-
-        async def subscribe(ws: Any) -> None:
-            await ws.subscribe_positions(update_interval=5.0)
-
-        async def on_message(msg: dict) -> None:
-            if msg.get("type") != "positions":
-                return
-            raw_data = msg.get("data", [])
-            # Update SDS cache
-            from condor.server_data_service import (
-                ServerDataType,
-                get_server_data_service,
-            )
-
-            get_server_data_service().put(
-                server_name, ServerDataType.POSITIONS, raw_data
-            )
-            await self._broadcast_update(channel, raw_data)
-
-        await self._run_ws_stream(
-            channel,
-            server_name,
-            label="Positions WS",
-            open_ws=lambda client: client.ws.executors(),
-            subscribe=subscribe,
-            on_message=on_message,
-        )
-
-    # -- Performance WS streaming (via Hummingbot /ws/executors performance) --
-
-    async def _performance_ws_stream(self, channel: str) -> None:
-        """Stream performance from Hummingbot /ws/executors and update SDS cache."""
-        parts = channel.split(":")
-        if len(parts) < 2:
-            return
-        server_name = parts[1]
-
-        async def subscribe(ws: Any) -> None:
-            await ws.subscribe_performance(update_interval=5.0)
-
-        async def on_message(msg: dict) -> None:
-            if msg.get("type") != "performance":
-                return
-            await self._broadcast_update(channel, msg.get("data", {}))
-
-        await self._run_ws_stream(
-            channel,
-            server_name,
-            label="Performance WS",
-            open_ws=lambda client: client.ws.executors(),
-            subscribe=subscribe,
-            on_message=on_message,
-        )
-
-    # -- Controller Performance polling stream --
-
-    async def _controller_perf_stream(self, channel: str) -> None:
-        """Poll latest controller performance every 30s and broadcast snapshots."""
-        parts = channel.split(":")
-        if len(parts) < 2:
-            return
-        server_name = parts[1]
-
-        from config_manager import get_config_manager
-
-        cm = get_config_manager()
-        backoff = 5
-
-        while True:
-            try:
-                if not self._has_subscribers(channel):
-                    logger.info(
-                        "No subscribers for %s, stopping controller perf stream",
-                        channel,
-                    )
-                    self._controller_perf_tasks.pop(channel, None)
-                    return
-
-                client = await cm.get_client(server_name)
-                result = (
-                    await client.bot_orchestration.get_latest_controller_performance()
-                )
-
-                # Normalize to list of snapshots
-                snapshots = []
-                if isinstance(result, list):
-                    snapshots = result
-                elif isinstance(result, dict):
-                    data = result.get(
-                        "data", result.get("snapshots", result.get("records", []))
-                    )
-                    if isinstance(data, list):
-                        snapshots = data
-                    elif isinstance(data, dict):
-                        for key, val in data.items():
-                            if isinstance(val, dict):
-                                val.setdefault("controller_id", key)
-                                snapshots.append(val)
-
-                if snapshots:
-                    await self._broadcast_update(channel, {"snapshots": snapshots})
-                    backoff = 5
-
-                await asyncio.sleep(30)
-
-            except asyncio.CancelledError:
-                return
-            except Exception as e:
-                logger.warning(
-                    "Controller perf stream error for %s: %s, retrying in %ds...",
-                    channel,
-                    e,
-                    backoff,
-                )
-                await asyncio.sleep(backoff)
-                backoff = min(backoff * 2, 120)
 
 
 # -- Singleton --
